Security Updates
@Zachary-Kuhn , I'm sorry to bring this to your concern. I see the following file is not supposed to be here. 
I don't know whether it as some use in Code.gov. But as far my knowledge it is a small human error where you saved your personnel file which everyone does. So I removed the file. If I understand wrongly and removed that pdf. Pardon me for it Mr.Zachry-Kuhn.