Deletion of Some file - Security
I exactly don't know, why the file is here. But I feel suspicious that this file contains some login credentials including the password. But I am not quite sure I'm right or wrong. Please remove this file as a security update and also change the password for the account. @RicardoAReyes if I'm made the wrong alert pardon me.