Increase the width of the Select tag
Select tag in the code.gov page is small which would give a jarring experience to the user. Update the style property inline to make it look better. 

Note : It is not good practice to embedded the style inline as I am not sure, the side effect of updating the value universally, I updated inline. Please have a look into it and update accordingly if need.